# backend/openapi.yml
openapi: 3.1.0
info:
  title: timerk
  version: 0.1.0
paths:
...
  /api/time-entries/start:
    post:
      tags:
      - time_entries
      summary: Start Timer
      operationId: start_timer_api_time_entries_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeEntryStartRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntryRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/time-entries:
    get:
      tags:
      - time_entries
      summary: List Time Entries
      operationId: list_time_entries_api_time_entries_get
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
          title: From
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
          title: To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeEntryRead'
                title: Response List Time Entries Api Time Entries Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/time-entries/{entry_id}:
    patch:
      tags:
      - time_entries
      summary: Update Entry
      operationId: update_entry_api_time_entries__entry_id__patch
      parameters:
      - name: entry_id
        in: path
        required: true
        schema:
          type: integer
          title: Entry Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeEntryUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntryRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - time_entries
      summary: Delete Entry
      operationId: delete_entry_api_time_entries__entry_id__delete
      parameters:
      - name: entry_id
        in: path
        required: true
        schema:
          type: integer
          title: Entry Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TimeEntryRead:
      properties:
        id:
          type: integer
          title: Id
        project_id:
          type: integer
          title: Project Id
        started_at:
          type: string
          format: date-time
          title: Started At
        stopped_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Stopped At
        duration_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Sec
      type: object
      required:
      - id
      - project_id
      - started_at
      title: TimeEntryRead
    TimeEntryStartRequest:
      properties:
        project_id:
          type: integer
          title: Project Id
      type: object
      required:
      - project_id
      title: TimeEntryStartRequest
    TimeEntryUpdateRequest:
      properties:
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        stopped_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Stopped At
      type: object
      title: TimeEntryUpdateRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError